policy user

+++ app/Http/Controllers/ThreadsController.php
@@ -122,6 +122,7 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
         $thread->delete();
 
         if (request()->wantsJson()) {
